WorkLess-Project/backend/requirements.txt
# FastAPI and Server
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6

# Authentication & Security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-dotenv==1.0.0

# Database
supabase==2.3.0

# AI & Document Processing
google-generativeai==0.3.1

# Data Validation
pydantic==2.5.0
pydantic-settings==2.1.0

# Utilities
aiofiles==23.2.1
pillow==10.1.0
pdf2image==1.16.3

# Logging & Monitoring
python-json-logger==2.0.7
